Unify errors under the CompilerError class. This way we can easily print all errors that are handled by our code and have error messages. Allows a real compiler experience.
